pyproject.toml
[project]
name = "patch2-0"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "flask-talisman>=1.1.0",
    "waitress>=3.0.0",
]
